refactor(quiz): change create and answer quiz to a single feature called quiz.
