--- lib/features/answer_quiz/data/datasources/answer_quiz_remote_datasource.dart
@@ -1,9 +1,0 @@
-import '../../../../core/platform/quiz_api_client.dart';
-
-abstract class AnswerQuizRemoteDataSourceInterface {}
-
-class AnswerQuizRemoteDataSource
-    implements AnswerQuizRemoteDataSourceInterface {
-  final QuizApiClient client;
-  AnswerQuizRemoteDataSource({required this.client});
-}
