--- lib/features/answer_quiz/data/repositories/answer_quiz_repository.dart
@@ -1,10 +1,0 @@
-import '../../domain/repositories/answer_quiz_repository_interface.dart';
-import '../datasources/answer_quiz_remote_datasource.dart';
-
-class AnswerQuizRepository implements AnswerQuizRepositoryInterface {
-  final AnswerQuizRemoteDataSourceInterface remoteDatasource;
-
-  AnswerQuizRepository({
-    required this.remoteDatasource,
-  });
-}
