--- lib/features/answer_quiz/domain/repositories/answer_quiz_repository_interface.dart
@@ -1,1 +1,0 @@
-abstract class AnswerQuizRepositoryInterface {}
