--- lib/features/create_quiz/domain/repositories/create_quiz_repository_interface.dart
@@ -1,6 +1,0 @@
-import '../../../../core/result/result.dart';
-import '../entities/question.dart';
-
-abstract class CreateQuizRepositoryInterface {
-  Future<IResult<List<Question>>> getQuestions();
-}
