+++ lib/features/quiz/domain/repositories/quiz_repository_interface.dart
@@ -1,0 +1,7 @@
+import '../../../../core/result/result.dart';
+import '../entities/question.dart';
+import '../usecases/get_questions.dart';
+
+abstract class QuizRepositoryInterface {
+  Future<IResult<List<Question>>> getQuestions(GetQuestionsParams params);
+}
